[FLINK-13028][table-api-java] Extract legacy type inference logic
